Isoplot

Documentation for Isoplot.jl.

Isoplot.UPbAnalysis

Isoplot.UPbAnalysis

Isoplot.yorkfit

Isoplot.UPbAnalysis — Type

struct UPbAnalysis{T<:AbstractFloat}

Core type for U-Pb analyses. Has fields

μ :: Vector{T<:AbstractFloat}
Σ :: Matrix{T<:AbstractFloat}

where μ contains the means

μ = [r²⁰⁷Pb²³⁵U, r²⁰⁶Pb²³⁸U]

and Σ contains the covariance matrix

Σ = [σ₇_₅^2 σ₇_₅*σ₃_₈
     σ₇_₅*σ₃_₈ σ₃_₈^2]

Isoplot.UPbAnalysis — Method

UPbAnalysis(r²⁰⁷Pb²³⁵U, σ²⁰⁷Pb²³⁵U, r²⁰⁶Pb²³⁸U, σ²⁰⁶Pb²³⁸U, correlation; T=Float64)

Construct a UPbAnalysis object from individual isotope ratios and (1-sigma!) uncertainties.

Examples

julia> UPbAnalysis(22.6602, 0.0175, 0.40864, 0.00017, 0.83183)
UPbAnalysis{Float64}([22.6602, 0.40864], [0.00030625000000000004 2.4746942500000003e-6; 2.4746942500000003e-6 2.8900000000000004e-8])

Isoplot.yorkfit — Method

yorkfit(x, σx, y, σy)

Uses the York (1968) two-dimensional least-squares fit to calculate a, b, and uncertanties σa, σb for the equation y = a + bx, given x, y and uncertaintes σx, `σy.

For further reference, see: York, Derek (1968) "Least squares fitting of a straight line with correlated errors" Earth and Planetary Science Letters 5, 320-324. doi: 10.1016/S0012-821X(68)80059-7

Examples

julia> x = (1:100) .+ randn.();

julia> y = 2*(1:100) .+ randn.();

julia> yorkfit(x, ones(100), y, ones(100))
YorkFit{Float64}:
Least-squares linear fit of the form y = a + bx where
  intercept a : -0.29 ± 0.2 (1σ)
  slope b     : 2.0072 ± 0.0035 (1σ)
  MSWD        : 0.8136665223891004